add system setup, binary bundler, airgap install
This commit makes numerous changes to support airgap. Chief among them are

* adding system_setup role
* adding binary_bundler role
* adding the taskfiles necessary to support installing with binaries

Breaking out system setup allows us to better control the order of operations that need to occur in a playbook for a redpanda cluster to be created in either a proxy or airgap configuration. It ensures that for proxy config we've properly handled the apt/dnf proxy before attempting to manage certs and installs, and for an airgap install it allows us to simply drop this step as unncessary.

binary bundler by contrast provides a quick and easy process for generating a bundle of the necessary RPMs/DEBs for a redpanda install given the necessary inputs (which are checked with asserts).

Finally the airgap install taskfiles in the broker role allow us to ensure that we are properly handling the copy and install of the DEB/RPM bundles, including the all important chowning to redpanda of the data directory for the dpkg install in particular.