add retry to issue certs
This task often fails due to a race condition with "Unable to rename /workdir/ansible/tls/ca/serial.txt.new to /workdir/ansible/tls/ca/serial.txt". While we could fix this task by setting serial to 1, this would render the playbook unusable as the number of brokers increases. Also, it usually doesn't hit the race condition. As an alternative solution to a really annoying but ultimately harmless bug I've added a retry with a small delay for errors.