handle console package
add console install to main role

added console config and tls

Doing this with tags in a playbook doesn't work. if you specify tags in a playbook when importing a role the tags element refers to the import role task, it isn't passed into the role and used to determine which tasks are run. Very annoying.

also added proxy changes back

also add a meta play so broker can be used just for installing certs