sasl and acl mgmt role rough pass
Concept here is that we use a map -- which could be jsonized and passed in through the CLI with a couple of changes -- to pass in users, passwords and permissions to rpk which then handles the process of create/delete.

I require two changes to two flags to delete users or permissions as I want to make it hard to do by accident. Deleting is bad mkay so it should be hard to do by accident.