op and base64 json vars
End users won't want to store their passwords in plaintext in yaml so I've added the ability to pass in a json object that has been base64 encoded. The intent here is to enable CI driven usage via a secure object store that can pass everything in thru the environment. I've opted to make this an override rather than a merge as I don't want to risk the example contaminating the end user's actual users. For desktop driven testing the yaml should be fine, but it's still an option to use the more secure route.

Operation is also made more generic to give users more control. I may just transform this into "pass in a string and we'll run it" as an alternative to the current approach if people prefer it.