feat: add require_client_auth to tls for mtls
This adds require_client_auth which will hopefully enable mtls support.

It also fixes the template test by guaranteeing it pulls from the defaults file. it also fixes an issue where pulling from the defaults file would result in unrendered expressions because of the chaining of variables in the defaults file. I fixed that by rendering the file twice and hitting it with some recursion just in case which fixed the issue of partial rendering of daisy chained variables from the defaults file.

I believe the technical term for this is "nuke it from orbit, it's the only way to be sure"