Setup correct fips packages when fips mode requested
RPMs have particular ordering quirks in the dependency chain so we
have to select the right package set and order when fips mode is on.
Without this, redpanda pulls in redpanda-rpk which conflicts with
the redpanda-rpk-fips package. So we need to pull in the rpk-fips
package first, and then the rest of the package chain.